feat: 피그마 기능 명세서 반영 (2.3)
